♻️ REFACTOR 상품 type으로 분기하여 화면에 보여줄 데이터를 만드는 custom hook을 구현하여 상품 아이템 컴포넌트 리팩토링
