+++ src/component/Common/CProductItem.tsx
@@ -9,27 +9,13 @@ import PreparingImage from "@asset/preparing-image.jpeg";
 import { AiFillStar } from "react-icons/ai";
 
 import { useToast } from "@hook/useToast";
+import { useProductInfo } from "@hook/useProductInfo";
 
-import { IProductInfo, IProductItemWithBookmark } from "@type/ProductList";
-import { PRODUCT_TYPE } from "@constant/ProductList";
+import { IProductItemWithBookmark } from "@type/ProductList";
 
 function CProductItem(props: { item: IProductItemWithBookmark }) {
-    const {
-        id,
-        title,
-        sub_title,
-        price,
-        image_url,
-        brand_name,
-        brand_image_url,
-        follower,
-        discountPercentage,
-        type,
-        isBookmarked,
-    } = props.item;
+    const { id, type, isBookmarked } = props.item;
 
-    // prettier-ignore
-    const [productInfo, setProductInfo] = useState<IProductInfo>({});
     const [isOpen, setIsOpen] = useState(false);
     const [bookmarked, setBookmarked] = useState(!!isBookmarked);
 
@@ -37,53 +23,11 @@ function CProductItem(props: { item: IProductItemWithBookmark }) {
     const addBookmarkFn = useSetRecoilState(addBookmark);
     const removeBookmarkFn = useSetRecoilState(removeBookmark);
 
-    const infoWidthCategory = () => {
-        if (type === PRODUCT_TYPE.PRODUCT) {
-            return {
-                title: title,
-                subTitle: sub_title,
-                info: `${discountPercentage || 0}%`,
-                subInfo: `${Number.parseInt(price || "0").toLocaleString()}원`,
-                img: image_url,
-            };
-        }
 
-        if (type === PRODUCT_TYPE.CATEGORY) {
-            return {
-                title: `#${title}`,
-                subTitle: "",
-                info: "",
-                subInfo: "",
-                img: image_url,
-            };
-        }
 
-        if (type === PRODUCT_TYPE.EXHIBITION) {
-            return {
-                title: title,
-                subTitle: "",
-                info: `${discountPercentage || 0}%`,
-                subInfo: "",
-                img: image_url,
-            };
-        }
 
-        if (type === PRODUCT_TYPE.BRAND) {
-            return {
-                title: brand_name,
-                subTitle: "",
-                info: "관심고객수",
-                subInfo: `${(follower || 0).toLocaleString()}명`,
-                img: brand_image_url,
-            };
         }
-
-        return {};
     };
-
-    useEffect(() => {
-        setProductInfo(infoWidthCategory());
-    }, []);
 
     return (
         <>
@@ -96,7 +40,7 @@ function CProductItem(props: { item: IProductItemWithBookmark }) {
                     <img
                         className="w-full h-full hover:scale-125 transition-all duration-400"
                         src={productInfo?.img || PreparingImage}
-                        alt={`${title} 이미지`}
+                        alt={`${productInfo?.title || ""} 이미지`}
                     />
                 </div>
                 <figcaption className="flex justify-between items-center h-6">
